+++ smoother.cc
@@ -9,13 +9,13 @@ namespace amg {
 //==============================================================================
 void gauss_seidel::apply_prev_smooth(const spmat_csr &A, const vec &rhs, vec &x, const vector<bool> *color_tag) const {
     const size_t row = A.rows();
-    for (size_t i = 0; i < row; ++i)
+    for (int64_t i = 0; i < row; ++i)
         iteration_body(A, rhs, x, i);
 }
 
 void gauss_seidel::apply_post_smooth(const spmat_csr &A, const vec &rhs, vec &x, const vector<bool> *color_tag) const {
     const size_t row = A.rows();
-    for (size_t i = row-1; i >= 0; --i)
+    for (int64_t i = row-1; i >= 0; --i)
         iteration_body(A, rhs, x, i);
 }
